get my nweets

+++ src/components/Nweet.js
@@ -8,6 +8,8 @@ const Nweet = ({ nweetObj, isOwner }) => {
     const ok = window.confirm("Are you sure you want to delete this nweet?");
     if (ok) {
       await dbService.doc(`nweets/${nweetObj.id}`).delete();
+    }
+    if (ok && nweetObj.attachmentUrl) {
       await storageService.refFromURL(nweetObj.attachmentUrl).delete();
     }
   };
